+++ libla/element_product_BENCHMARK.cc
@@ -18,10 +18,10 @@ class DenseMatrixElementProductBench :
     public Benchmark
 {
     private:
-        int _size;
+        unsigned long _size;
         int _count;
     public:
-        DenseMatrixElementProductBench(const std::string & id, int size, int count) :
+        DenseMatrixElementProductBench(const std::string & id, unsigned long size, int count) :
             Benchmark(id)
         {
             _size = size;
