+++ libla/sum_BENCHMARK.cc
@@ -19,11 +19,11 @@ class ScalarDenseMatrixSumBench :
     public Benchmark
 {
     private:
-        int _size;
+        unsigned long _size;
         int _count;
 
     public:
-        ScalarDenseMatrixSumBench(const std::string & id, int size, int count) :
+        ScalarDenseMatrixSumBench(const std::string & id, unsigned long size, int count) :
             Benchmark(id)
         {
             _size  = size;
@@ -56,10 +56,10 @@ class DenseVectorSumBench :
     public Benchmark
 {
     private:
-        int _size;
+        unsigned long _size;
         int _count;
     public:
-        DenseVectorSumBench(const std::string & id, int size, int count) :
+        DenseVectorSumBench(const std::string & id, unsigned long size, int count) :
             Benchmark(id)
         {
             _size = size;
